test/integration_tests/requirements.txt
dotenv
pandas>=2.0
pyarrow>=14.0
pyiceberg[rest]>=0.6.0
requests>=2.31.0
findspark>=2.0.1
pyspark==3.5.1
pytest==8.4.1
snowflake-connector-python>=3.10.0